requirements.txt
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
glpi-api==0.5.0
idna==3.7
pycparser==2.22
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
urllib3==2.2.2
